extract_upload_logs/requirements.txt
requests
google-cloud-bigquery
google-cloud-secret-manager